fix: 修复 zoom level 类型错误

+++ include/QCefView.h
@@ -294,7 +294,7 @@ public:
   /// Get the zoom level value.
   /// </summary>
   /// <returns>the zoom level value.</returns>
-  int zoomLevel();
+  double zoomLevel();
 
 signals:
   /// <summary>
